requirements.txt
click
psycopg2-binary
python-daemon
service
paramiko
boto3
humanize
